# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards Management System API
  description: |
    API для управления банковскими картами, пользователями и транзакциями.
    Включает аутентификацию, работу с картами, переводы средств и административные функции.
  version: 1.0.0
  contact:
    name: API Support
    email: support@bankcards.com
  license:
    name: Proprietary
    url: https://www.bankcards.com/terms
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
...
paths:
...
  /api/transactions/transfer:
    post:
      tags:
        - Transactions
      summary: Перевод между картами
      description: Перевод средств между своими картами
      operationId: transferBetweenCards
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '201':
          description: Перевод успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Ошибка валидации или недостаточно средств
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/transactions:
    get:
      tags:
        - Transactions
      summary: Получить транзакции пользователя
      description: Получить историю транзакций текущего пользователя
      operationId: getUserTransactions
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Номер страницы
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Размер страницы
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Список транзакций
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTransactionResponse'
  /api/transactions/card/{cardId}:
    get:
      tags:
        - Transactions
      summary: Получить транзакции карты
      description: Получить историю транзакций конкретной карты
      operationId: getCardTransactions
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Транзакции карты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTransactionResponse'
  /api/transactions/{transactionId}:
    get:
      tags:
        - Transactions
      summary: Получить транзакцию по ID
      description: Получить детали конкретной транзакции
      operationId: getTransaction
      security:
        - bearerAuth: []
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Детали транзакции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Транзакция не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/transactions/card/{cardId}/daily-limit:
    get:
      tags:
        - Transactions
      summary: Проверить дневной лимит
      description: Проверить использованный дневной лимит для карты
      operationId: checkDailyLimit
      security:
        - bearerAuth: []
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о дневном лимите
          content:
            application/json:
              example:
                success: true
                message: Daily limit checked successfully
                data:
                  dailyTotal: 1500.50
                  dailyLimit: 5000.00
                  remaining: 3499.50
                  limitExceeded: false
...
components:
  schemas:
...
    TransferRequest:
      type: object
      required:
        - fromCardNumber
        - toCardNumber
        - amount
        - cvv
      properties:
        fromCardNumber:
          type: string
          pattern: '^\d{16}$'
          example: '4149471805568597'
        toCardNumber:
          type: string
          pattern: '^\d{16}$'
          example: '4149476684550065'
        amount:
          type: number
          format: decimal
          minimum: 0.01
          maximum: 1000000
          example: 500.50
        description:
          type: string
          maxLength: 255
          example: Перевод между моими картами
        cvv:
          type: string
          pattern: '^\d{3}$'
          example: '244'
    TransactionResponse:
      type: object
      properties:
        transactionId:
          type: string
          example: 'txn_7s8df9s7d8f9s'
        amount:
          type: number
          format: decimal
          example: 500.50
        currency:
          type: string
          example: RUB
        type:
          type: string
          example: TRANSFER
        status:
          type: string
          example: COMPLETED
        description:
          type: string
          example: Перевод между картами
        fromCardMasked:
          type: string
          example: '**** **** **** 8597'
        toCardMasked:
          type: string
          example: '**** **** **** 0065'
        transactionDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    # Response обертки
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
        data:
          type: object
          description: Данные ответа
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message
        error:
          type: string
          example: Detailed error description
        timestamp:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00'
        path:
          type: string
          example: '/api/cards'
    PaginatedResponse:
      type: object
      properties:
        content:
          type: array
          items:
            type: object
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          example: 150
        totalPages:
          type: integer
          example: 8
...
    PaginatedTransactionResponse:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/TransactionResponse'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен аутентификации
security:
  - bearerAuth: []